requirements.txt
pygame==2.1.2
grpcio==1.51.1
grpcio-tools==1.51.1
